Add more fields for subscription
This adds firstName, lastName and salutation to the registration. Fields are integrated in TCA, DB, Fluid and Scheduler.
The salutation MUST be a field "SALUTATION" in mailchimp with 2 possible values "Herr" or "Frau" (German for male or female salutation).
Update your database in install tool to add the new fields.