all fields are required, styled for parsley (powermail like)
Now all fields are required. Note that in mailchimp the salutation is a custom merge_field. Added constants to constant editor for success and failed Pid. This is needed if you insert the plugin via Typoscript.
Also adds error labels in EN and DE.
FIx wrong array index in condition in Form->createAction()